+++ man-roxygen/param_learner_properties.R
@@ -4,6 +4,7 @@
 #'   The following properties are currently standardized and understood by learners in \CRANpkg{mlr3}:
 #'   * `"missings"`: The learner can handle missing values in the data.
 #'   * `"weights"`: The learner supports observation weights.
+#'   * `"offset"`: The learner can incorporate offset values to adjust predictions.
 #'   * `"importance"`: The learner supports extraction of importance scores, i.e. comes with an `$importance()` extractor function (see section on optional extractors in [Learner]).
 #'   * `"selected_features"`: The learner supports extraction of the set of selected features, i.e. comes with a `$selected_features()` extractor function (see section on optional extractors in [Learner]).
 #'   * `"oob_error"`: The learner supports extraction of estimated out of bag error, i.e. comes with a `oob_error()` extractor function (see section on optional extractors in [Learner]).