+++ NEWS.md
@@ -1,5 +1,7 @@
 # mlr3 (development version)
 
+* feat: add new `col_role` offset in `Task` and offset `Learner` property.
+A warning is produced if a learner that doesn't support offsets is trained with a task that has an offset column.
 * fix: the `$predict_newdata()` method of `Learner` now automatically conducts type conversions (#685)
 * BREAKING_CHANGE: Predicting on a `task` with the wrong column information is now an error and not a warning.
 * Column names with UTF-8 characters are now allowed by default.